requirementsp37.txt
absl-py==1.4.0
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
aniso8601==9.0.1
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2023.5.7
charset-normalizer==3.1.0
clang==5.0
click==8.1.3
elastic-transport==8.4.0
elasticsearch==7.8.0
et-xmlfile==1.1.0
Flask==2.0.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
flatbuffers==23.5.26
frozenlist==1.3.3
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.54.2
h5py==3.1.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.0.1
Jinja2==3.1.2
joblib==1.2.0
keras==2.11.0
Keras-Preprocessing==1.1.2
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
multidict==6.0.4
mysql-connector-python==8.0.33
nltk==3.6.7
numpy==1.21.6
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
pandas==1.3.5
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.7.0
python-dateutil==2.8.2
python-pcre==0.7
pytz==2023.3
redis==4.5.5
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.15.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==1.1.0
tqdm==4.65.0
twilio==8.2.2
typing-extensions==3.7.4.3
urllib3==1.26.16
Werkzeug==2.2.3
wrapt==1.12.1
yarl==1.9.2
zipp==3.15.0
